setting basic swap
