+++ .env.example
@@ -1,0 +1,4 @@
+PRIVATE_KEY=0000000000000000000000000000000000000000000000000000000000000000
+BASE_MAINNET_URL=https://mainnet.base.org/
+API_KEY=00000000
+BASE_URL=https://dextra-node-00000000000.app/v2/quote
