+++ .gitignore
@@ -1,5 +1,6 @@
 # Logs
 logs
+.idea
 *.log
 npm-debug.log*
 yarn-debug.log*
